+++ basicswap/js_server.py
@@ -551,10 +551,17 @@ def js_getcoinseed(self, url_split, post_string, is_json) -> bytes:
         raise ValueError('Particl wallet seed is set from the Basicswap mnemonic.')
 
     ci = swap_client.ci(coin)
-    seed = swap_client.getWalletKey(coin, 1)
+    if coin == Coins.XMR:
+        key_view = swap_client.getWalletKey(coin, 1, for_ed25519=True)
+        key_spend = swap_client.getWalletKey(coin, 2, for_ed25519=True)
+        address = ci.getAddressFromKeys(key_view, key_spend)
+        return bytes(json.dumps({'coin': ci.ticker(), 'key_view': ci.encodeKey(key_view), 'key_spend': ci.encodeKey(key_spend), 'address': address}), 'UTF-8')
+
+    seed_key = swap_client.getWalletKey(coin, 1)
     if coin == Coins.DASH:
-        return bytes(json.dumps({'coin': ci.ticker(), 'seed': seed.hex(), 'mnemonic': ci.seedToMnemonic(seed)}), 'UTF-8')
-    return bytes(json.dumps({'coin': ci.ticker(), 'seed': seed.hex()}), 'UTF-8')
+        return bytes(json.dumps({'coin': ci.ticker(), 'seed': seed_key.hex(), 'mnemonic': ci.seedToMnemonic(seed_key)}), 'UTF-8')
+    seed_id = ci.getSeedHash(seed_key)
+    return bytes(json.dumps({'coin': ci.ticker(), 'seed': seed_key.hex(), 'seed_id': seed_id.hex()}), 'UTF-8')
 
 
 def js_setpassword(self, url_split, post_string, is_json) -> bytes: