+++ doc/release-notes.md
@@ -17,7 +17,9 @@
 - api: Can abandon bids.
 - If wallets are encrypted the system will only load in-progress bids when unlocked rather than at startup.
 - Can set overrides for automation strategies per identity.
-- ui: bids on expired offers won't show as available
+- ui: Bids on expired offers won't show as available.
+- api: getcoinseed shows seed id.
+- ui: Can edit automation strategy data.
 
 
 0.0.54